divide h_mearsure.h into paired header files
