+++ raster/r.texture/matvec.c
@@ -5,7 +5,29 @@
 #include <grass/raster.h>
 #include <grass/glocale.h>
 
-#include "h_measure.h"
+#include "matvec.h"
+
+int bsearch_gray(int *array, int n, int val)
+{
+    int lo, hi, mid;
+
+    lo = 0;
+    hi = n - 1;
+
+    while (lo <= hi) {
+        mid = (lo + hi) >> 1;
+
+        if (array[mid] == val)
+            return mid;
+
+        if (array[mid] > val)
+            hi = mid - 1;
+        else
+            lo = mid + 1;
+    }
+
+    return -1;
+}
 
 float *vector(int n)
 {
